# [XABT] Break BuildApk into individual tasks for each content type.