… src/Xamarin.Android.Build.Tasks/Tasks/CollectJarContentFilesForArchive.cs
… src/Xamarin.Android.Build.Tasks/Tasks/CollectNativeFilesForArchive.cs
… src/Xamarin.Android.Build.Tasks/Tasks/CollectRuntimeConfigFilesForArchive.cs
… src/Xamarin.Android.Build.Tasks/Tasks/CollectTypeMapFilesForArchive.cs
… src/Xamarin.Android.Build.Tasks/Utilities/PackageFileListBuilder.cs
… src/Xamarin.Android.Build.Tasks/Xamarin.Android.Common.targets